Progress check point.
Added a short limit feature.
For now, the short ratio appears on the market page, and a new pre-screen method on CallMarket adjusts orders to ensure we are still within the limit.

Next, I'll need to tie that into the process, and run tests that it works the way I expect.